+++ morib/src/main/java/org/morib/server/domain/allowedGroup/application/DeleteAllowedGroupService.java
@@ -1,0 +1,6 @@
+package org.morib.server.domain.allowedGroup.application;
+
+public interface DeleteAllowedGroupService {
+    void deleteAllowedGroupById(Long groupId);
+
+}
